pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "addresskr"
authors = [{name = "Pak Youngrok", email = "youngrok@finecode.kr"}]
license = {file = "LICENSE"}
classifiers = ["License :: OSI Approved :: Apache Software License"]
dynamic = ["version", "description"]
dependencies = [
    "requests",
    "typer",
    "ipython",
    "python-dotenv",
]

[project.urls]
Home = "https://github.com/finecodekr/addresskr"
